processor/requirements.txt
rembg[cpu]
mediapipe
pillow
fastapi
uvicorn
python-multipart
